# docs/swagger.yaml
definitions:
...
  filesystem.QuotaInfo:
    properties:
      files:
        type: integer
      files_grace:
        type: string
      files_limit:
        type: integer
      files_quota:
        type: integer
      filesystem:
        type: string
      kbytes:
        type: integer
      kbytes_grace:
        type: string
      kbytes_limit:
        type: integer
      kbytes_quota:
        type: integer
    type: object
...
  models.SacctRequest:
    properties:
      accounts:
        items:
          type: string
        type: array
      allclusters:
        type: boolean
      allusers:
        type: boolean
      arrayjobs:
        type: boolean
      brief:
        type: boolean
      clusters:
        items:
          type: string
        type: array
      completion:
        type: boolean
      duplicates:
        type: boolean
      endtime:
        type: string
      format:
        type: string
      jobids:
        items:
          type: string
        type: array
      jobnames:
        items:
          type: string
        type: array
      long:
        type: boolean
      maxcpus:
        type: integer
      maxnodes:
        type: integer
      mincpus:
        type: integer
      minnodes:
        type: integer
      nodelist:
        items:
          type: string
        type: array
      noheader:
        type: boolean
      parsable:
        type: boolean
      partitions:
        items:
          type: string
        type: array
      qos:
        items:
          type: string
        type: array
      starttime:
        type: string
      states:
        items:
          type: string
        type: array
      truncate:
        type: boolean
      users:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/filesystem/quota/:
    get:
      consumes:
      - application/json
      description: 获取指定用户或文件系统的磁盘配额信息，包括已使用空间、配额限制、文件数量等详细信息
      parameters:
      - description: SSH会话密钥
        example: '"tsh_a2e932b625c0d598db3800aa91b92016"'
        in: header
        name: sessionKey
        required: true
        type: string
      - description: 集群名称
        example: '"hpc1"'
        in: query
        name: cluster
        required: true
        type: string
      - description: 查询路径
        example: '"/home/user"'
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取配额信息成功" example({"filesystem":"/dev/sda1","kbytes":1024000,"kbytes_quota":2048000,"kbytes_limit":2097152,"kbytes_grace":"none","files":1000,"files_quota":5000,"files_limit":10000,"files_grace":"none"})
          schema:
            $ref: '#/definitions/filesystem.QuotaInfo'
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: 路径不存在或配额信息不可用
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误或用户未登录
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 获取磁盘配额信息
      tags:
      - 文件管理
  /api/v1/slurm/sacct/accounting/:
    post:
      consumes:
      - application/json
      description: 获取作业会计信息，包括作业ID、用户、状态、提交时间等详细信息
      parameters:
      - description: sacct请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SacctRequest'
      - description: SSH会话密钥
        example: '"tsh_a2e932b625c0d598db3800aa91b92016"'
        in: header
        name: sessionKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功，返回作业会计信息列表
          schema:
            properties:
              home_path:
                type: string
              session_key:
                type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: 认证失败，用户名或密码错误
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误或SSH连接失败
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 获取作业会计信息
      tags:
      - 作业管理
  /api/v1/slurm/sacct/job/user/{user}/:
    post:
      consumes:
      - application/json
      description: 获取作业会计信息，包括作业ID、用户、状态、提交时间等详细信息
      parameters:
      - description: sacct请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SacctRequest'
      - description: SSH会话密钥
        example: '"tsh_a2e932b625c0d598db3800aa91b92016"'
        in: header
        name: sessionKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功，返回作业会计信息列表
          schema:
            properties:
              home_path:
                type: string
              session_key:
                type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: 认证失败，用户名或密码错误
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误或SSH连接失败
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 获取指定用户作业会计信息
      tags:
      - 作业管理
  /api/v1/slurm/sacct/jobs/:
    post:
      consumes:
      - application/json
      description: 获取作业会计信息，包括作业ID、用户、状态、提交时间等详细信息
      parameters:
      - description: sacct请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SacctRequest'
      - description: SSH会话密钥
        example: '"tsh_a2e932b625c0d598db3800aa91b92016"'
        in: header
        name: sessionKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功，返回作业会计信息列表
          schema:
            properties:
              home_path:
                type: string
              session_key:
                type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: 认证失败，用户名或密码错误
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误或SSH连接失败
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 获取作业会计信息
      tags:
      - 作业管理
  /api/v1/slurm/sacct/jobs/{jobid}/:
    post:
      consumes:
      - application/json
      description: 获取作业会计信息，包括作业ID、用户、状态、提交时间等详细信息
      parameters:
      - description: sacct请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SacctRequest'
      - description: SSH会话密钥
        example: '"tsh_a2e932b625c0d598db3800aa91b92016"'
        in: header
        name: sessionKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功，返回作业会计信息列表
          schema:
            properties:
              home_path:
                type: string
              session_key:
                type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: 认证失败，用户名或密码错误
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误或SSH连接失败
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 获取指定作业会计信息
      tags:
      - 作业管理
...
swagger: "2.0"